requirements.txt
python-dotenv
riotwatcher